Add first draft of tree-sitter face
Not super happy with how these turned out. I tried to be as close to
"vanilla" as possible, but TS has the "feature" that it applies
multiple faces to each token, which means that they often overlap in
strange ways. Not really sure how to deal with that right now, but I
do know that my TSX and CSS look really stupid.